Move js to separate file
