+++ app/javascript/application.js
@@ -1,16 +1,4 @@
 import { initAll } from "govuk-frontend";
+import "./cookie_banner"
 
 initAll();
-
-// Hide cookie banner
-var cookieBanner = document.querySelector(".govuk-cookie-banner");
-if (cookieBanner) {
-  var hideButton = cookieBanner.querySelector('.cookie-hide-button');
-
-  if (hideButton) {
-    hideButton.addEventListener("click", function(e) {
-      e.preventDefault();
-      cookieBanner.style.display = 'none';
-    });
-  }
-}
